change Intro to IntroPage and start styling
