--- src/scss/_modal.scss
@@ -1,289 +1,0 @@
-// Modals
-
-//Show the homepage modal
-.intro-show {
-  font-size: 0.9rem;
-  position: absolute;
-  z-index: 1030;
-  top: 1.9rem;
-  right: 0;
-
-  input {
-    margin-right: .35rem;
-  }
-
-  @include media-breakpoint-up(sm) {
-    z-index: 1030;
-    top: 4.75rem;
-    right: 0;
-  }
-
-  @include media-breakpoint-up(md) {
-    z-index: 1030;
-    top: 1.85rem;
-    right: 1rem;
-  }
-}
-
-//Homepage Intro Modal
-.intro {
-  font-size: 1rem;
-
-  .intro-bg {
-    object-fit: contain;
-    position: relative;
-  }
-
-  &.modal.show {
-    .modal-dialog {
-      max-width: 90%;
-      max-height: 90%;
-      margin-left: auto;
-      margin-right: auto;
-    }
-  }
-
-  .modal-content {
-    background-image:
-            url('/assets/images/act-1-mobile-top.png'),
-            url('/assets/images/act-1-mobile-bottom.png'),
-            url('/assets/images/act-1-mobile-stretch.png');
-    background-position:
-            top center,
-            bottom center,
-            top center;
-    background-repeat: no-repeat;
-    background-size:
-            contain,
-            contain,
-            cover;
-
-    @include media-breakpoint-up(lg) {
-      background: transparent;
-    }
-  }
-
-  .modal-header {
-    position: absolute;
-    z-index: 10;
-    top: 0;
-    right: 0;
-    border: 0;
-
-    .btn-close {
-      @include btn-close-white();
-    }
-  }
-
-  .modal-body {
-    border: 0;
-    color: #000;
-    margin: 27% 10% 12% 2%;
-
-    p {
-      margin-bottom: 0.25rem;
-      line-height: 1.1;
-    }
-  }
-
-  /*
-  .modal-dialog {
-  background: transparent url('../assets/images/act-1.png') no-repeat center top;
-  background-size: contain;
-  }
-  .modal-body {
-    padding: 4rem;
-    background-image:
-            url('../assets/icons/flourish-top-left.svg'),
-            url('../assets/icons/flourish-top-right.svg'),
-            url('../assets/icons/flourish-bottom-left.svg'),
-            url('../assets/icons/flourish-bottom-right.svg');
-    background-position:
-            top 10px left 10px,
-            top 10px right 10px,
-            bottom 10px left 10px,
-            bottom 10px right 10px;
-    background-repeat: no-repeat;
-
-    @include color-mode(light) {
-      svg {
-        fill: $purple-400;
-      }
-    }
-  } */
-
-  .arkham-choice {
-    margin-top: 1rem;
-
-    p {
-      margin-bottom: 0.5rem;
-    }
-
-    a {
-      font-weight: 700;
-      color: #6c1f25;
-
-      &:hover,
-      &:focus,
-      &:active {
-        color: #a4363c;
-        border-bottom: 3px solid #a4363c;
-      }
-    }
-  }
-
-  .modal-footer {
-    margin-top: -5%;
-    z-index: 10;
-    padding-right: 4%;
-    padding-bottom: 0.25rem;
-    font-weight: 600;
-    border: 0;
-
-    .form-check {
-      color: #ffffff;
-
-      .form-check-input {
-        border-color: #fff;
-      }
-
-      &:hover,
-      &:focus,
-      &:active {
-        border-width: 2px;
-      }
-    }
-
-    .form-check-input:checked {
-      background-color: #a4363c;
-      border-color: #a4363c;
-    }
-  }
-
-  @include media-breakpoint-up(sm) {
-    font-size: 2.75vw;
-
-    .modal-body {
-      margin: 28% 12% 12% 2%;
-
-      p {
-        margin-bottom: 0.4rem;
-        //line-height: 1.3;
-      }
-    }
-
-    .arkham-choice {
-      p {
-        margin-bottom: 0.25rem;
-      }
-    }
-
-    .modal-footer {
-      margin-top: -7.5%;
-    }
-  }
-
-  @include media-breakpoint-up(md) {
-    .modal-dialog {
-      max-width: 90%;
-    }
-
-    .modal-body {
-      margin: 30% 13% 12% 2%;
-
-      p {
-        margin-bottom: 0.5rem;
-        //line-height: 1.3;
-        line-height: 1.15;
-      }
-    }
-
-    .arkham-choice {
-      p {
-        //margin-bottom: 0.35rem;
-        margin-bottom: 0.25rem;
-      }
-
-      a {
-        text-underline-offset: 6px;
-      }
-    }
-
-    .modal-footer {
-      //margin-top: -5.5%;
-      margin-top: -7.5%;
-    }
-  }
-
-  @include media-breakpoint-up(lg) {
-    //font-size: 1.4vw;
-    font-size: 1.525vw;
-
-    .modal-dialog {
-      max-width: 90%;
-      border: 1px solid $gray-700;
-    }
-
-    .modal-body {
-      position: absolute;
-      top: 0;
-      left: 0;
-      width: 50%;
-      margin: 17% 0 10% 2.5%;
-
-      p {
-        //line-height: 1.3;
-      }
-    }
-
-    .arkham-choice {
-      margin-top: 1rem;
-
-      p {
-        margin-bottom: 0.3rem;
-      }
-    }
-
-    .modal-footer {
-      margin-top: -5%;
-      padding-right: 43%;
-    }
-  }
-
-  @include media-breakpoint-up(xl) {
-    .modal-dialog {
-      max-width: 90%;
-    }
-
-    .modal-body {
-      margin: 18% 0 10% 3%;
-
-      p {
-        //line-height: 1.4;
-      }
-    }
-
-    .modal-footer {
-      margin-top: -4.5%;
-    }
-  }
-
-  @include media-breakpoint-up(xxl) {
-    .modal-dialog {
-      max-width: 90%;
-    }
-
-    .modal-body {
-      margin: 18% 0 10% 3%;
-
-      p {
-        margin-bottom: 0.5rem;
-        //line-height: 1.4;
-      }
-    }
-
-    .modal-footer {
-      margin-top: -4%;
-    }
-  }
-}
