+++ src/scss/styles.scss
@@ -114,4 +114,3 @@ $font-weight-base: 300;
 @import "_projects.scss";
 @import "_accordion.scss";
 @import "_buttons.scss";
-@import "_modal.scss";
